--- next.config.js
@@ -1,13 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-    experimental: {
-        appDir: false,
-    },
-    webpack(config) {
-        config.module.rules.push({ test: /\.svg$/, use: ['@svgr/webpack'] });
-
-        return config;
-    },
-};
-
-module.exports = nextConfig
