+++ next.config.mjs
@@ -1,0 +1,13 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  experimental: {
+    appDir: false,
+  },
+  webpack(config) {
+    config.module.rules.push({ test: /\.svg$/, use: ['@svgr/webpack'] });
+
+    return config;
+  },
+};
+
+export default nextConfig;
