+++ package.json
@@ -1,39 +1,40 @@
 {
-    "name": "portfolio",
-    "version": "0.1.0",
-    "private": true,
-    "scripts": {
-        "dev": "next dev",
-        "build": "next build",
-        "start": "next start",
-        "lint": "next lint"
-    },
-    "dependencies": {
-        "@chakra-ui/react": "^2.4.9",
-        "@emotion/react": "^11.10.5",
-        "@emotion/styled": "^11.10.5",
-        "@next/font": "13.1.6",
-        "@types/node": "18.13.0",
-        "@types/react": "18.0.27",
-        "@types/react-dom": "18.0.10",
-        "@vercel/speed-insights": "^1.1.0",
-        "eslint": "^9.15.0",
-        "eslint-config-next": "13.1.6",
-        "framer-motion": "^9.0.2",
-        "graphql": "^16.9.0",
-        "graphql-request": "^7.1.2",
-        "next": "13.1.6",
-        "react": "18.2.0",
-        "react-dom": "18.2.0",
-        "swr": "^2.2.5",
-        "typescript": "4.9.5"
-    },
-    "devDependencies": {
-        "@eslint/js": "^9.15.0",
-        "@svgr/webpack": "^6.5.1",
-        "eslint-config-sznm": "^2.0.4",
-        "eslint-plugin-react": "^7.37.2",
-        "globals": "^15.12.0",
-        "typescript-eslint": "^8.14.0"
-    }
+  "name": "portfolio",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
+  },
+  "dependencies": {
+    "@chakra-ui/react": "^2.4.9",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@next/font": "13.1.6",
+    "@types/node": "18.13.0",
+    "@types/react": "18.0.27",
+    "@types/react-dom": "18.0.10",
+    "@vercel/speed-insights": "^1.1.0",
+    "eslint": "^9.15.0",
+    "eslint-config-next": "13.1.6",
+    "framer-motion": "^9.0.2",
+    "graphql": "^16.9.0",
+    "graphql-request": "^7.1.2",
+    "next": "13.1.6",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "swr": "^2.2.5",
+    "typescript": "4.9.5"
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.15.0",
+    "@svgr/webpack": "^6.5.1",
+    "eslint-config-sznm": "^2.0.4",
+    "eslint-plugin-react": "^7.37.2",
+    "globals": "^15.12.0",
+    "typescript-eslint": "^8.14.0"
+  }
 }
